refactor(draw): introduce olDraw.composable to simplify and concentrate logic
