+++ src/composables/map/ol.composable.ts
@@ -1,8 +1,5 @@
 import type BaseLayer from 'ol/layer/Base'
 import type OlMap from 'ol/Map'
-import VectorLayer from 'ol/layer/Vector'
-import VectorSource from 'ol/source/Vector'
-import { DrawnFeature } from '@/services/draw/drawn-feature'
 
 import { layersCache } from '@/stores/layers.cache'
 import type { Layer, LayerId } from '@/stores/map.store.model'
@@ -12,7 +9,6 @@ import { statePersistorStyleService } from '@/services/state-persistor/state-per
 import { olLayerFactoryService } from '@/services/ol-layer/ol-layer-factory.service'
 import { olLayerTimeService } from '@/services/ol-layer/ol-layer-time.service'
 import { OlLayer } from '@/services/ol-layer/ol-layer.model'
-import useMap from './map.composable'
 
 const DEFAULT_BGZINDEX = -200 // Value comming  from legacy
 
@@ -150,21 +146,6 @@ export default function useOpenLayers() {
     }
   }
 
-  function setDrawnFeatures(features: DrawnFeature[], olMap: OlMap) {
-    const mapLayers = olMap.getLayers()
-    const drawLayer = mapLayers
-      .getArray()
-      .find(
-        layer => layer.get('featureID') === 'featureLayer'
-      ) as VectorLayer<VectorSource>
-    if (!drawLayer) {
-      useMap().addDrawLayer()
-    } else {
-      drawLayer.getSource()?.clear()
-      drawLayer.getSource()?.addFeatures(features)
-    }
-  }
-
   return {
     addLayer,
     findLayer,
@@ -176,6 +157,5 @@ export default function useOpenLayers() {
     getLayerFromCache,
     setBgLayer,
     applyOnBgLayer,
-    setDrawnFeatures,
   }
 }
